refactor: extract server net code to a Client class
This separates the net code handling from the Player class to a Client class, making the giant mess of player.ts a bit less messy...

And will also allow joining a game directly as spectator in the future without hacks

...also refactors spectating because old code was garbage :)
