+++ shared/net/spectateMsg.ts
@@ -1,22 +1,20 @@
 import type { AbstractMsg, BitStream } from "./net.ts";
 
+export enum SpectateAction {
+    None,
+    Begin,
+    Next,
+    Prev,
+}
+
 export class SpectateMsg implements AbstractMsg {
-    specBegin = false;
-    specNext = false;
-    specPrev = false;
-    specForce = false;
+    action: SpectateAction = SpectateAction.None;
 
     serialize(s: BitStream) {
-        s.writeBoolean(this.specBegin);
-        s.writeBoolean(this.specNext);
-        s.writeBoolean(this.specPrev);
-        s.writeBoolean(this.specForce);
+        s.writeUint8(this.action);
     }
 
     deserialize(s: BitStream) {
-        this.specBegin = s.readBoolean();
-        this.specNext = s.readBoolean();
-        this.specPrev = s.readBoolean();
-        this.specForce = s.readBoolean();
+        this.action = s.readUint8();
     }
 }
